requirements.txt
faiss-cpu
sentence-transformers
gradio
python-docx
transformers
requests